+++ iti-react/Changelog.md
@@ -168,4 +168,5 @@ Then:
 
 # 2.3.5 [wip]
 
-- Go back to using `@use-it/event-listener` instead of my fork
+- \*\*\* Go back to using `@use-it/event-listener` instead of my fork
+- Bring in `usePopoverClickListener`